Greatly optimised vertex animations and vertex animation compiler and massively improved fragment shader Will look into how to continue improving the fragment shader and the lighting in general

+++ Vertex_Buffer_Declarations.h
@@ -11,7 +11,7 @@ public:
 	unsigned int Buffer_ID = Unassigned_Bit_Mask;
 	unsigned int Vertex_Array_ID = Unassigned_Bit_Mask;
 
-	uint16_t Buffer_Storage_Hint = GL_STATIC_DRAW;
+	size_t Buffer_Storage_Hint = GL_STATIC_DRAW;
 
 	// Don't necessarily need index buffers
